fix: 불필요한 코드 삭제, DTO 분리
